add stride
